+++ code/Model/Indexer/Algoliaadditionalsections.php
@@ -5,8 +5,8 @@ class Algolia_Algoliasearch_Model_Indexer_Algoliaadditionalsections extends Mage
     const EVENT_MATCH_RESULT_KEY = 'algoliasearch_match_result';
 
     /** @var Algolia_Algoliasearch_Model_Resource_Engine */
-    private $engine;
-    private $config;
+    protected $engine;
+    protected $config;
 
     public function __construct()
     {